Cleanup tests and stabilize.
