+++ tests/test_steps_mp.py
@@ -1,12 +1,56 @@
-# ActivitySim
-# See full license in LICENSE.txt.
-import subprocess
-from tests.data_hash import hash_dataframe
 from pathlib import Path
 import pandas as pd
+from tests.data_hash import hash_dataframe
 
+from activitysim.core import config
+from activitysim.core import tracing
+from activitysim.core import pipeline
 from activitysim.core import inject
+from activitysim.core import mp_tasks
 
+from populationsim import steps
+
+TAZ_COUNT = 36
+TAZ_100_HH_COUNT = 33
+TAZ_100_HH_REPOP_COUNT = 26
+
+
+def setup_dirs():
+
+    example_dir = Path(__file__).parent.parent / 'examples'
+    example_configs_dir = (example_dir / 'example_test' / 'configs').__str__()
+    configs_dir = (Path(__file__).parent / 'configs').__str__()    
+    mp_configs_dir = (example_dir / 'example_test' / 'configs_mp').__str__()
+    output_dir = Path(__file__).parent / 'output'
+    data_dir = (example_dir / 'example_test' / 'data').__str__()
+    
+
+    inject.add_injectable("configs_dir", [mp_configs_dir, configs_dir, example_configs_dir])
+    inject.add_injectable("output_dir", output_dir)
+    inject.add_injectable("data_dir", data_dir)
+
+    tracing.config_logger()
+
+    tracing.delete_output_files("csv")
+    tracing.delete_output_files("txt")
+    tracing.delete_output_files("yaml")
+
+
+def regress():
+
+    expanded_household_ids = pipeline.get_table("expanded_household_ids")
+    assert isinstance(expanded_household_ids, pd.DataFrame)
+    taz_hh_counts = expanded_household_ids.groupby("TAZ").size()
+    assert len(taz_hh_counts) == TAZ_COUNT
+    assert taz_hh_counts.loc[100] == TAZ_100_HH_COUNT
+
+    # output_tables action: skip
+    output_dir = inject.get_injectable("output_dir")
+    assert not (output_dir / 'households.csv').exists()
+    assert (output_dir / 'summary_DISTRICT_1.csv').exists()
+    
+    result_hash = hash_dataframe(expanded_household_ids, sort_by = ['hh_id', 'TRACT', 'TAZ', 'PUMA'])
+    assert result_hash == '05d7f8d0bf5d8e5c7ee29b67c13d858f'
 
 def teardown_function(func):
     inject.clear_cache()
@@ -15,16 +59,26 @@ def teardown_function(func):
 
 def test_mp_run():
 
-    file_path = Path(__file__).parent / 'run_mp.py'
-    output_dir = Path(__file__).parent / 'output'
+    setup_dirs()
 
-    subprocess.check_call(['coverage', 'run', file_path])
-    
-    # This hash is the md5 of the dataframe string file previously generated
-    # by the pipeline. It is used to check that the pipeline is generating the same output.
-    expanded_household_ids = pd.read_csv(output_dir / 'expanded_household_ids.csv')
-    assert hash_dataframe(expanded_household_ids) == '37b263bfa2d25c48fac9c591a87d91df'
+    # Debugging ----------------------
+    run_list = mp_tasks.get_run_list()
+    mp_tasks.print_run_list(run_list)
+    # --------------------------------
 
-if __name__ == '__main__':
+    # do this after config.handle_standard_args, as command line args
+    # may override injectables
+    injectables = ["data_dir", "configs_dir", "output_dir"]
+    injectables = {k: inject.get_injectable(k) for k in injectables}
+
+    mp_tasks.run_multiprocess(injectables)
+
+    pipeline.open_pipeline("_")
+    regress()
+    pipeline.close_pipeline()
+
+
+
+if __name__ == "__main__":
 
     test_mp_run()
